pyproject.toml
[project]
name = "edurag"
version = "0.1.0"
description = "Full workflow for generation edurag from Stepik courses"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate==1.12.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asyncio==4.0.0",
    "audioread==3.1.0",
    "av==16.0.1",
    "beautifulsoup4==4.14.3",
    "bitsandbytes==0.49.0",
    "boto3==1.42.16",
    "botocore==1.42.16",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "ctranslate2==4.6.1",
    "decorator==5.2.1",
    "dotenv==0.9.9",
    "fastapi==0.122.0",
    "faster-whisper==1.2.1",
    "filelock==3.19.1",
    "flatbuffers==25.9.23",
    "fsspec==2025.9.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.28.1",
    "google-api-python-client==2.187.0",
    "google-auth==2.43.0",
    "google-auth-httplib2==0.2.1",
    "google-genai==1.52.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "h2==4.3.0",
    "hf-xet==1.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "joblib==1.5.2",
    "jsonschema>=4.25.1",
    "lazy-loader==0.4",
    "librosa==0.11.0",
    "llama-cpp-python>=0.3.16",
    "llvmlite==0.45.1",
    "markupsafe==2.1.5",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "networkx==3.5",
    "numba==0.62.1",
    "numpy==2.2.6",
    "onnxruntime==1.23.2",
    "opencv-python==4.12.0.88",
    "packaging==25.0",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "pooch==1.8.2",
    "portalocker==3.2.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.1.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pydub==0.25.1",
    "pyparsing==3.2.5",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pywin32==311",
    "pyyaml==6.0.3",
    "qdrant-client==1.16.2",
    "regex==2025.11.3",
    "requests==2.32.5",
    "rsa==4.9.1",
    "s3transfer==0.16.0",
    "safetensors==0.7.0",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "sentence-transformers==5.2.0",
    "services>=0.1.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "soupsieve==2.8",
    "soxr==1.0.0",
    "starlette==0.50.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "torch==2.10.0.dev20251124+cu128",
    "torchaudio==2.10.0.dev20251124+cu128",
    "torchvision==0.25.0.dev20251124+cu128",
    "tqdm==4.67.1",
    "transformers==4.57.3",
    "typer-slim==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "websockets==15.0.1",
    "wheel==0.45.1",
]
[[tool.uv.index]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-nightly" }
torchvision = { index = "pytorch-nightly" }
torchaudio = { index = "pytorch-nightly" }

[tool.uv]
environments = [
    "sys_platform == 'win32'"
]
